Changed the proposal.
After diving in to the project for some time, I realized just how
different ProLog was from the other languages I have used. In ProLog,
there is no such thing as time, I knew this, but the implications were
what escaped me.

After the experience I gained, I figured out that it would be very
difficult to have an actual roster print, rather than on a student by
student basis. So I changed that to the "A" level deliverable, and
balanced out the other goals.