Hourly Commit and C - Level
Got the D and C level done in this same hour, but this C level can also
be considered an improvement on the D level, because it actually has a
function to return the list, rather than requiring the user to make an
advanced query.

use roster(College_ID) to see the students that would make it into the
college.